+++ src/main/java/org/example/game_elements/Rogue.java
@@ -20,7 +20,7 @@ public class Rogue extends Hero {
                 new HeroAttribute(1, 4, 1));
     }
 
-    public int calculateDamage(){
+    public float calculateDamage(){
         return 0;
     }
 }
